+++ Pooptube/Client/Tree2.cpp
@@ -78,7 +78,7 @@ void Tree2::_SetBillBoardTexture()
 	float angle = pooptube::CalculateAngle( GetFrontVector(), mBillBoard->GetFrontVector() );
 	angle /= D3DX_PI;
 	angle *= 4;
-	int dir( angle );
+	int dir( static_cast<int>(angle) );
 	dir += 8;
 	dir %= 8;
 	switch( dir ){
